split gradle plugin common to allow sharing code between gradle webapp plugins
