+++ azure-functions-gradle-plugin/src/main/java/com/microsoft/azure/plugin/functions/gradle/handler/DeployHandler.java
@@ -9,9 +9,9 @@ import com.azure.core.management.AzureEnvironment;
 import com.azure.core.management.exception.ManagementException;
 import com.google.common.base.Preconditions;
 import com.microsoft.azure.functions.annotation.AuthorizationLevel;
+import com.microsoft.azure.gradle.configuration.GradleRuntimeConfig;
+import com.microsoft.azure.gradle.temeletry.TelemetryAgent;
 import com.microsoft.azure.plugin.functions.gradle.IAppServiceContext;
-import com.microsoft.azure.plugin.functions.gradle.configuration.GradleRuntimeConfig;
-import com.microsoft.azure.plugin.functions.gradle.telemetry.TelemetryAgent;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.applicationinsights.ApplicationInsights;
 import com.microsoft.azure.toolkit.lib.applicationinsights.ApplicationInsightsEntity;
