+++ azure-functions-gradle-plugin/src/main/java/com/microsoft/azure/plugin/functions/gradle/task/DeployTask.java
@@ -5,12 +5,11 @@
 package com.microsoft.azure.plugin.functions.gradle.task;
 
 
+import com.microsoft.azure.gradle.temeletry.TelemetryAgent;
+import com.microsoft.azure.gradle.util.GradleProxyUtils;
 import com.microsoft.azure.plugin.functions.gradle.AzureFunctionsExtension;
 import com.microsoft.azure.plugin.functions.gradle.GradleFunctionContext;
 import com.microsoft.azure.plugin.functions.gradle.handler.DeployHandler;
-import com.microsoft.azure.plugin.functions.gradle.telemetry.TelemetryAgent;
-
-import com.microsoft.azure.plugin.functions.gradle.util.GradleProxyUtils;
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
