+++ azure-functions-gradle-plugin/src/main/java/com/microsoft/azure/plugin/functions/gradle/task/PackageZipTask.java
@@ -4,12 +4,11 @@
  */
 package com.microsoft.azure.plugin.functions.gradle.task;
 
+import com.microsoft.azure.gradle.temeletry.TelemetryAgent;
 import com.microsoft.azure.plugin.functions.gradle.AzureFunctionsExtension;
 import com.microsoft.azure.plugin.functions.gradle.GradleFunctionContext;
 import com.microsoft.azure.plugin.functions.gradle.handler.PackageHandler;
-import com.microsoft.azure.plugin.functions.gradle.telemetry.TelemetryAgent;
 import com.microsoft.azure.plugin.functions.gradle.util.FunctionUtils;
-
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
@@ -18,7 +17,6 @@ import org.gradle.api.tasks.TaskAction;
 import org.zeroturnaround.zip.ZipUtil;
 
 import javax.annotation.Nullable;
-
 import java.io.File;
 import java.io.IOException;
