+++ azure-gradle-plugins-common/src/main/java/com/microsoft/azure/gradle/configuration/GradleRuntimeConfig.java
@@ -2,7 +2,8 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
-package com.microsoft.azure.plugin.functions.gradle.configuration;
+
+package com.microsoft.azure.gradle.configuration;
 
 import lombok.Getter;
 import lombok.Setter;
